+++ ci/do_ci.sh
@@ -82,6 +82,14 @@ function bazel_binary_build() {
 }
 
 if [[ "$1" == "bazel.release" ]]; then
+  # When testing memory consumption, we want to test against exact byte-counts
+  # where possible. As these differ between platforms and compile options, we
+  # define the 'release' builds as canonical and test them only in CI, so the
+  # toolchain is kept consistent. This ifdef is checked in
+  # test/common/stats/stat_test_utility.cc when computing
+  # Stats::TestUtil::MemoryTest::mode().
+  BAZEL_BUILD_OPTIONS="${BAZEL_BUILD_OPTIONS} --cxxopt=-DMEMORY_TEST_EXACT=1"
+
   setup_clang_toolchain
   echo "bazel release build with tests..."
   bazel_binary_build release