+++ source/common/config/resources.h
@@ -20,6 +20,7 @@ public:
   const std::string VirtualHost{"type.googleapis.com/envoy.api.v2.route.VirtualHost"};
   const std::string ScopedRouteConfiguration{
       "type.googleapis.com/envoy.api.v2.ScopedRouteConfiguration"};
+  const std::string Runtime{"type.googleapis.com/envoy.service.discovery.v2.Runtime"};
 };
 
 using TypeUrl = ConstSingleton<TypeUrlValues>;