+++ source/common/stats/isolated_store_impl.h
@@ -10,6 +10,7 @@
 
 #include "common/common/utility.h"
 #include "common/stats/heap_stat_data.h"
+#include "common/stats/null_gauge.h"
 #include "common/stats/store_impl.h"
 #include "common/stats/symbol_table_impl.h"
 #include "common/stats/utility.h"